update for ^3.2
change admin route to admin/plugins/rss-widget, so it doesn't conflict with nodebb-plugin-rss and matches other plugin pages
moved admin.js to public/admin.js
updated templates to use bootstrap5 classes
