+++ package.json
@@ -17,7 +17,7 @@
   "main": "plugin.js",
   "name": "nodebb-widget-rss",
   "nbbpm": {
-    "compatibility": "^1.18.0"
+    "compatibility": "^3.2.0"
   },
   "repository": {
     "type": "git",
