+++ templates/admin/plugins/rss-widget.tpl
@@ -1,0 +1,18 @@
+<div class="acp-page-container">
+	<!-- IMPORT admin/partials/settings/header.tpl -->
+  <div class="row m-0">
+    <div id="spy-container" class="col-12 px-0 mb-4" tabindex="0">
+      <form class="rss-settings" role="form">
+        <p class="lead">A widget that shows your rss feeds</p>
+        <div class="mb-3">
+          <label class="form-label" for="urls">Urls</label>
+          <input class="form-control" id="urls" name="urls" placeholder="Comma separated list of urls to RSS feeds." title="Urls" type="text">
+        </div>
+        <div class="">
+          <label class="form-label" for="limit">limit</label>
+          <input class="form-control" id="limit" name="limit" placeholder="Number entries form each feed url. Default is 10." title="Limit" type="text">
+        </div>
+      </form>
+    </div>
+  </div>
+</div>
